fix(remote-files): fallback to get when remote server not support head method (#10370)
